Restructure part 1
