+++ src/include/dipp_error.h
@@ -1,3 +1,9 @@
+#ifndef DIPP_ERROR_H
+#define DIPP_ERROR_H
+
+#include "dipp_paramids.h"
+#include "vmem_storage.h"
+#include <param/param.h>
 #include <stdlib.h>
 #include <stdint.h>
 
@@ -37,3 +43,9 @@ extern uint8_t err_current_pipeline;
 extern uint8_t err_current_module;
 
 uint32_t get_error_as_uint32(ERROR_CODE code);
+void set_error_param(ERROR_CODE error_code);
+
+/* Define error log parameters */
+PARAM_DEFINE_STATIC_VMEM(PARAMID_ERROR_LOG, log_status, PARAM_TYPE_UINT32, -1, 0, PM_CONF, NULL, NULL, storage, VMEM_ERROR_CODE, "Latest error code");
+
+#endif
