+++ src/include/dipp_main.h
@@ -1,0 +1,43 @@
+#ifndef DIPP_MAIN_H
+#define DIPP_MAIN_H
+
+#include "dipp_paramids.h"
+#include "vmem_storage.h"
+#include <param/param.h>
+#include <stdlib.h>
+#include <stdint.h>
+
+#define MSG_QUEUE_KEY 71
+
+// Return codes
+#define SUCCESS 0
+#define FAILURE -1
+
+// Pipeline run codes
+typedef enum
+{
+    PROCESS_STOP = 0,
+    PROCESS_ONE = 1,
+    PROCESS_ALL = 2,
+    PROCESS_WAIT_ONE = 3,
+    PROCESS_WAIT_ALL = 4
+} PIPELINE_PROCESS;
+
+typedef struct ImageBatch {
+    long mtype;          /* message type to read from the message queue */
+    int height;          /* height of images */
+    int width;           /* width of images */
+    int channels;        /* channels of images */
+    int num_images;      /* amount of images */
+    int batch_size;      /* size of the image batch */
+    int shm_key;         /* key to shared memory segment of image data */
+    int pipeline_id;     /* id of pipeline to utilize for processing */
+    unsigned char *data; /* address to image data (in shared memory) */
+} ImageBatch;
+
+/* Define a pipeline_run parameter */
+static uint8_t _pipeline_run = 0;
+void callback_run(param_t *param, int index);
+PARAM_DEFINE_STATIC_RAM(PARAMID_PIPELINE_RUN, pipeline_run, PARAM_TYPE_UINT8, -1, 0, PM_CONF, callback_run, NULL, &_pipeline_run, "Set the pipeline to execute the file");
+
+#endif
