+++ src/include/stb_image_write.h
