+++ src/include/vmem_storage.h
